bench-resource-optimizer/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9
langchain==0.2.5
langchain-openai==0.1.9
langchain-community==0.2.5
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
PyPDF2==3.0.1
pydantic==2.7.4
python-dotenv==1.0.1
sentence-transformers==3.0.1
